# Gestion_Escolar/Microservicio_Asistencia/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Attendance Management API",
      "version": "1.0.0",
      "title": "Attendance Management"
    },
    "host": "localhost:3006",
    "basePath": "/api",
    "tags": [
      {
        "name": "attendance",
        "description": "Attendance management"
      }
    ],
    "paths": {
      "/attendance": {
        "get": {
          "tags": ["attendance"],
          "summary": "Get all attendances",
          "description": "",
          "operationId": "getAttendances",
          "consumes": ["application/json"],
          "produces": ["application/json"],
          "parameters": [],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Attendance"
                }
              }
            }
          }
        },
        "post": {
          "tags": ["attendance"],
          "summary": "Create a new attendance",
          "description": "",
          "operationId": "createAttendance",
          "consumes": ["application/json"],
          "produces": ["application/json"],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Attendance object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Attendance"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Attendance created",
              "schema": {
                "$ref": "#/definitions/Attendance"
              }
            }
          }
        }
      },
      "/attendance/{id}": {
        "put": {
          "tags": ["attendance"],
          "summary": "Update an attendance",
          "description": "",
          "operationId": "updateAttendance",
          "consumes": ["application/json"],
          "produces": ["application/json"],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Attendance ID",
              "required": true,
              "type": "integer",
              "format": "int64"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Attendance object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Attendance"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Attendance updated"
            }
          }
        }
      }
    },
    "definitions": {
      "Attendance": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "estudiante_id": {
            "type": "integer",
            "format": "int64"
          },
          "curso_id": {
            "type": "integer",
            "format": "int64"
          },
          "fecha": {
            "type": "string",
            "format": "date"
          },
          "asistencia": {
            "type": "boolean"
          }
        }
      }
    }
  }
  